fix: Build errors with vue-qrcode-reader
