--- src/aux-server/aux-web/shared/public/vue-qrcode-reader/index.js
@@ -1,4 +1,0 @@
-// Forked from https://github.com/gruhn/vue-qrcode-reader
-export { default } from './dist/vue-qrcode-reader.common'
-export * from './dist/vue-qrcode-reader.common'
-import './dist/vue-qrcode-reader.css'
