+++ src/aux-server/aux-web/shared/public/vue-qrcode-reader/src/misc/errors.ts
