+++ src/aux-server/aux-web/shared/public/vue-qrcode-reader/src/mixins/CommonAPI.ts
@@ -1,0 +1,17 @@
+export default {
+    methods: {
+      async onDetect(resultPromise: any) {
+        this.$emit("detect", resultPromise);
+  
+        try {
+          const { content } = await resultPromise;
+  
+          if (content !== null) {
+            this.$emit("decode", content);
+          }
+        } catch (error) {
+          // fail silently
+        }
+      }
+    }
+  };
